requirements.txt
Flask-Session===0.3.1
Flask==1.1.1
mysql-connector==2.2.9
python-dotenv==0.12.0
flask_restful==0.3.8
Werkzeug==0.16.0